Video Search: Change to direct-mapped radix search
The HammingTree search "tree" was effectively the same
as a radix search that set the radix dynamically.

Now use a pure radix search and the user can set the radix,
which also determines the accuracy.

Also adds vradix parameter for changing the radix

Also removes a feature for matching short videos as it was
extremely slow.